refactor: move list sdk and init bid sdk
