+++ src/constants.ts
@@ -1,3 +1,4 @@
 import { PublicKey } from "@solana/web3.js";
 
-export const SOUNDWORK_PROGRAM_ID = new PublicKey("EUmBNHvFqhkA6Uaqv6pDup5ESHKCqoAweQ4kzAMjNZhX");
+export const SOUNDWORK_LIST_PROGRAM_ID = new PublicKey("EUmBNHvFqhkA6Uaqv6pDup5ESHKCqoAweQ4kzAMjNZhX");
+export const SOUNDWORK_BID_PROGRAM_ID = new PublicKey("CVuoapcC1RG8y1m86eChkXmynPi4FaykDC8jM8soMZ4j");
