+++ src/soundwork-list/main.ts
@@ -1,0 +1,1 @@
+export * from './sdk';
